Now can start

+++ app/Controllers/FantaCISFController.php
@@ -7,7 +7,7 @@ use Lepton\Authenticator\UserAuthenticator;
 use Lepton\Controller\BaseController;
 use Liquid\{Liquid, Template};
 
-use App\Models\{User, FantaCISFMember, FantaCISFTeam, FantaCISFPoints};
+use App\Models\{User, FantaCISFMember, FantaCISFTeam, FantaCISFPoints, FantaCISFSettings};
 use Lepton\Authenticator\LoginRequired;
 use Lepton\Http\HttpResponse;
 
@@ -56,49 +56,53 @@ class FantaCISFController extends BaseController
     #[LoginRequired(level: 1)]
     public function index()
     {
-        return $this->render("FantaCISF/index");
+        $setting = FantaCISFSettings::get(name: "has_started");
+        return $this->render("FantaCISF/index", ["can_edit" => 1-$setting->value]);
     }
 
 
     #[LoginRequired(1)]
     public function toggle($id)
     {
-        $user = (new UserAuthenticator())->getLoggedUser();
-        $member = FantaCISFMember::get($id);
-        $price = 0;
-        switch($member->role) {
-            case 1:
-                $price = 20;
-                break;
-            case 2:
-                $price = 10;
-                break;
-            case 3:
-                $price = 6;
-                break;
-            default:
-                $price = 0;
-        }
-        $exists = FantaCISFTeam::filter(user: $user, teamMember: $member)->count();
-        $team_number = FantaCISFTeam::filter(user: $user)->count();
-        if($exists) {
-            FantaCISFTeam::get(user: $user, teamMember: $member)->delete();
-            echo "cancello";
-            $user->fantacisf_budget += $price;
-            $user->save();
-        } else {
-            if(($user->fantacisf_budget >= $price) && ($team_number < 5)) {
-                $teamAssociation = FantaCISFTeam::new(user: $user, teamMember: $member);
-                $teamAssociation->save();
-                $user->fantacisf_budget -= $price;
+        $setting = FantaCISFSettings::get(name: "has_started");
+        if($setting->value == 0) {
+            $user = (new UserAuthenticator())->getLoggedUser();
+            $member = FantaCISFMember::get($id);
+            $price = 0;
+            switch($member->role) {
+                case 1:
+                    $price = 20;
+                    break;
+                case 2:
+                    $price = 10;
+                    break;
+                case 3:
+                    $price = 6;
+                    break;
+                default:
+                    $price = 0;
+            }
+            $exists = FantaCISFTeam::filter(user: $user, teamMember: $member)->count();
+            $team_number = FantaCISFTeam::filter(user: $user)->count();
+            if($exists) {
+                FantaCISFTeam::get(user: $user, teamMember: $member)->delete();
+                echo "cancello";
+                $user->fantacisf_budget += $price;
                 $user->save();
             } else {
-                exit;
-            }
+                if(($user->fantacisf_budget >= $price) && ($team_number < 5)) {
+                    $teamAssociation = FantaCISFTeam::new(user: $user, teamMember: $member);
+                    $teamAssociation->save();
+                    $user->fantacisf_budget -= $price;
+                    $user->save();
+                } else {
+                    exit;
+                }
 
+            }
+            return new HttpResponse(200);
         }
         return new HttpResponse(200);
-
 
     }
 
@@ -106,17 +110,23 @@ class FantaCISFController extends BaseController
     #[LoginRequired(1)]
     public function saveName()
     {
-        $user = (new UserAuthenticator())->getLoggedUser();
-        $user->fantacisf_team = $_POST["team_name"];
-        $user->save();
+        $setting = FantaCISFSettings::get(name: "has_started");
+        if($setting->value == 0) {
+            $user = (new UserAuthenticator())->getLoggedUser();
+            $user->fantacisf_team = $_POST["team_name"];
+            $user->save();
 
-        return $this->showTeam(1);
+            return $this->showTeam(1);
+        } else {
+            return new HttpResponse(201, body: "");
+        }
     }
 
 
 
     #[LoginRequired(1)]
-    public function showTeam($update){
+    public function showTeam($update)
+    {
         $user = (new UserAuthenticator())->getLoggedUser();
         $membriEC = FantaCISFMember::filter(role: 1);
         $membriLC = FantaCISFMember::filter(role: 2);
@@ -127,15 +137,18 @@ class FantaCISFController extends BaseController
             $selected[] = $member->teamMember->id;
         }
 
+        $setting = FantaCISFSettings::get(name: "has_started");
+
         $data = [ "EC" => $membriEC,
         "LC" => $membriLC,
         "OC" => $membriOC,
         "selected" => $selected,
-        "user" => $user];
+        "user" => $user,
+        "can_edit" => 1-$setting->value];
 
         $headers = array();
 
-        if($update == 1){
+        if($update == 1) {
             $data["show_toast"] = true;
             $data["is_update"] = true;
             $headers["HX-Trigger"] = "showToast";
@@ -144,7 +157,8 @@ class FantaCISFController extends BaseController
     }
 
 
-    public static function cmp($a, $b){
+    public static function cmp($a, $b)
+    {
         return $a["points"] < $b["points"];
     }
 
@@ -154,26 +168,50 @@ class FantaCISFController extends BaseController
         $users = User::filter(fantacisf_team__neq: "");
         $standings = array();
 
-        foreach($users as $user){
+        foreach($users as $user) {
+            $points = $this->computePointsUser($user);
             $standings[] = [
                 "id" => $user->id,
                 "name" => $user->name." ".$user->surname,
                 "team_name" => $user->fantacisf_team,
-                "points" => $this->computePointsUser($user),
-                "team" => FantaCISFTeam::filter(user: $user)->do()
+                "points" => $points,
+                "team" => FantaCISFTeam::filter(user: $user)->do(),
+                "position" => 0
             ];
+
         }
+
+        $last_points = -1;
+        $last_position = 0;
+
         usort($standings, array(self::class, "cmp"));
+
+        foreach($standings as $key => &$standing){
+            if($key == 0 ){
+                $standing["position"] = 1;
+            } else if($last_points > $standing["points"]){
+                $standing["position"] = $last_position + 1;
+            } else {
+                $standing["position"] = $last_position;
+            }
+
+
+            $last_points = $standing["points"];
+            $last_position = $standing["position"];
+        }
+
+        $setting = FantaCISFSettings::get(name:"has_started");
         return $this->render("FantaCISF/league", ["users" => $standings,
-        "num_teams" => $users->count()]);
+        "num_teams" => $users->count(), "has_started" => $setting->value]);
     }
 
 
     #[LoginRequired(1)]
-    public function computePointsUser($user){
+    public function computePointsUser($user)
+    {
         $team = FantaCISFTeam::filter(user: $user);
         $points = 0;
-        foreach($team as $member){
+        foreach($team as $member) {
             $points += $this->computePointsMember($member->teamMember);
         }
         return $points;
@@ -182,10 +220,11 @@ class FantaCISFController extends BaseController
 
 
     #[LoginRequired(1)]
-    public function computePointsMember($member){
+    public function computePointsMember($member)
+    {
         $memberbonus = FantaCISFPoints::filter(member: $member);
         $points = 0;
-        foreach($memberbonus as $bonus){
+        foreach($memberbonus as $bonus) {
             $points += $bonus->bonus->points*$bonus->multiplier;
         }
         return $points;
